Add routes for different language pages
