+++ src/pages/home-en.tsx
@@ -1,0 +1,5 @@
+import { I18nWrapper } from "@/I18nWrapper";
+
+export default function Component() {
+  return <I18nWrapper language="en" />;
+}
